refactor: split portable session internals
